feat(comment): Add comment creation and management endpoints
This commit introduces endpoints for creating, updating, and
deleting comments.  It also includes improved input validationand authorization checks to ensure data integrity and
security. The CommentRequestDTO and CommentResponseDTO were
created to better handle data transfer. Swagger documentation was added to the endpoints.

+++ src/main/java/br/com/soupaulodev/forumhub/modules/comment/usecase/DeleteCommentUsecase.java
@@ -3,25 +3,42 @@ package br.com.soupaulodev.forumhub.modules.comment.usecase;
 import br.com.soupaulodev.forumhub.modules.comment.entity.CommentEntity;
 import br.com.soupaulodev.forumhub.modules.comment.repository.CommentRepository;
 import br.com.soupaulodev.forumhub.modules.exception.usecase.CommentNotFoundException;
+import br.com.soupaulodev.forumhub.modules.exception.usecase.TopicNotFoundException;
+import br.com.soupaulodev.forumhub.modules.exception.usecase.UnauthorizedException;
+import br.com.soupaulodev.forumhub.modules.exception.usecase.UserNotFoundException;
+import br.com.soupaulodev.forumhub.modules.topic.entity.TopicEntity;
+import br.com.soupaulodev.forumhub.modules.topic.repository.TopicRepository;
+import br.com.soupaulodev.forumhub.modules.user.entity.UserEntity;
+import br.com.soupaulodev.forumhub.modules.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.UUID;
 
 /**
  * Service class for deleting comments.
+ *
+ * @author <a href="https://soupaulodev.com.br">soupauldev</a>
  */
 @Service
 public class DeleteCommentUsecase {
 
-    public final CommentRepository commentRepository;
+    private final CommentRepository commentRepository;
+    private final UserRepository userRepository;
+    private final TopicRepository topicRepository;
 
     /**
      * Constructs a new DeleteCommentUsecase with the specified repository.
      *
      * @param commentRepository the repository for managing comments
+     * @param userRepository the repository for managing users
+     * @param topicRepository the repository for managing topics
      */
-    public DeleteCommentUsecase(CommentRepository commentRepository) {
+    public DeleteCommentUsecase(CommentRepository commentRepository,
+                                UserRepository userRepository,
+                                TopicRepository topicRepository) {
         this.commentRepository = commentRepository;
+        this.userRepository = userRepository;
+        this.topicRepository = topicRepository;
     }
 
     /**
@@ -29,11 +46,28 @@ public class DeleteCommentUsecase {
      *
      * @param id the UUID of the comment to be deleted
      * @throws CommentNotFoundException if the comment is not found
+     * @throws UserNotFoundException if the user or topic is not found
+     * @throws UnauthorizedException if the user is not allowed to delete a comment for another user or topic
+     * @throws TopicNotFoundException if the topic is not found
      */
-    public void execute(UUID id) {
+    public void execute(UUID id, UUID getAuthenticatedUserId) {
 
         CommentEntity commentFound = commentRepository.findById(id)
                 .orElseThrow(CommentNotFoundException::new);
+
+        UserEntity user = userRepository.findById(commentFound.getUser().getId())
+                .orElseThrow(UserNotFoundException::new);
+        if (!user.getId().equals(getAuthenticatedUserId)) {
+            throw new UnauthorizedException("You are not allowed to update a comment for another user");
+        }
+
+        TopicEntity topic = topicRepository.findById(commentFound.getTopic().getId())
+                .orElseThrow(TopicNotFoundException::new);
+
+        if (!user.participateInForum(topic.getForum())) {
+            throw new UnauthorizedException("You are not allowed to update a comment for this topic");
+        }
+
         commentRepository.delete(commentFound);
     }
 }
